python-service/requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
python-multipart==0.0.9
PyMuPDF==1.24.7
python-docx==1.1.2
spacy==3.7.5
pydantic==2.8.2
